Extract RoutesBodyGenerator from RoutesGenerator
